Split server and getting information about the browser
